--- src/TypealizR/Builder/TypeModel.cs
@@ -1,21 +1,0 @@
-﻿using Microsoft.CodeAnalysis;
-
-namespace TypealizR.Builder;
-public class TypeModel
-{
-	public TypeModel(string @namespace, string name, Accessibility accessibility = Accessibility.Internal)
-	{
-		Namespace = @namespace;
-		Name = name;
-        Accessibility = accessibility;
-	}
-
-    public string Namespace { get; }
-    public string Name { get; }
-
-	public Accessibility Accessibility { get; }
-
-	public string FullName => $"{Namespace}.{Name}";    public string GlobalFullName => $"global::{FullName}";    public string FullNameForClassName => FullName.Replace(".", "");
-
-
-}
