+++ src/TypealizR/Core/IMemberModel.cs
