+++ src/TypealizR/StringFormatterSourceGenerator/StringFormatterClassBuilder.cs
