+++ src/TypealizR/StringLocalizerExtensionsSourceGenerator/ExtensionClassBuilder.cs
