+++ src/TypealizR/StringLocalizerExtensionsSourceGenerator/ExtensionClassModel.cs
