+++ src/TypealizR/TypealizedClassSourceGenerator/InstanceMemberBuilder.cs
