+++ src/TypealizR/TypealizedClassSourceGenerator/InstanceMemberModel.cs
