+++ src/TypealizR/TypealizedClassSourceGenerator/TypealizedClassBuilder.cs
