+++ src/TypealizR/TypealizedClassSourceGenerator/TypealizedClassModel.cs
